museum_flask_app/requirements.txt
APScheduler==3.10.4
asn1crypto==1.5.1
binary==1.0.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mirakuru==2.5.1
mock==5.1.0
packaging==23.2
pg8000==1.30.2
pika==1.3.2
pluggy==1.3.0
pool==0.1.2.dev0
port-for==0.7.2
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
psutil==5.9.5
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-postgresql==5.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.21
testing.common.database==2.0.3
testing.postgresql==1.3.0
typing_extensions==4.8.0
tzlocal==5.1
urllib3==2.0.6
Werkzeug==2.3.7
